app/backend/requirements.txt
azure-identity==1.13.0b3
Flask==2.2.5
langchain==0.0.187
openai==0.26.4
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
pypdf==3.5.0
langchain==0.0.187
pypdf==3.5.0